+++ cloudwanderer/__init__.py
@@ -1,10 +1,12 @@
 """I wandered lonely through the cloud."""
 from . import storage_connectors
+from . import cloud_wanderer_resource
 from .cloud_wanderer import CloudWanderer
 from .aws_urn import AwsUrn
 from .custom_resource_definitions import CustomResourceDefinitions
 __all__ = [
     'storage_connectors',
+    'cloud_wanderer_resource',
     'CloudWanderer',
     'AwsUrn',
     'CustomResourceDefinitions'