+++ cloudwanderer/cloud_wanderer.py
@@ -1,6 +1,5 @@
 """Main cloudwanderer module."""
-from typing import Callable, List
-from collections import namedtuple
+from typing import List
 import logging
 from typing import TYPE_CHECKING
 import concurrent.futures
@@ -191,7 +190,7 @@ class CloudWanderer():
             client_args (dict): Arguments to pass into the boto3 client.
                 See: :meth:`boto3.session.Session.client`
         """
-        logger.info('Writing resource attributes in all regions')
+        logger.info('Writing resource attributes in all enabled regions')
         for region_name in self.enabled_regions:
             self.write_resource_attributes_in_region(
                 region_name=region_name,
@@ -329,78 +328,3 @@ class CloudWanderer():
             resource_type=xform_name(resource.meta.resource_model.shape),
             resource_id=resource_id
         )
-
-
-class ResourceMetadata(namedtuple('ResourceMetadata', ['resource_data', 'resource_attributes'])):
-    """Metadata for a :class:`CloudWandererResource`.
-
-    Contains the original dictionaries of the resource and its attributes.
-
-    Attributes:
-        resource_data (dict): The raw dictionary representation of the Resource.
-        resource_attributes (list): the list of raw dictionary representation of the Resource's attributes.
-    """
-
-
-class CloudWandererResource():
-    """A simple representation of a resource that prevents any storage metadata polluting the resource dictionary.
-
-    Use ``dict(my_resource_dict)`` to convert this object into a dictionary that
-    contains *only* the resource's metadata.
-
-    Attributes:
-        urn (cloudwanderer.aws_urn.AwsUrn): The AWS URN of the resource.
-        metadata (dict): The original storage representation of the resource as it was passed in.
-    """
-
-    def __init__(self, urn: AwsUrn, resource_data: dict,
-                 resource_attributes: List[dict] = None, loader: Callable = None) -> None:
-        """Initialise the resource."""
-        self.urn = urn
-        self.cloudwanderer_metadata = ResourceMetadata(
-            resource_data=resource_data,
-            resource_attributes=resource_attributes or []
-        )
-        self._loader = loader
-        self._set_attrs()
-
-    def load(self) -> None:
-        """Inflate this resource with all data from the original storage connector it was spawned from.."""
-        if self._loader is None:
-            logger.error('Could not inflate %s, storage connector loader not populated', self)
-            return
-        updated_resource = self._loader(urn=self.urn)
-        if updated_resource is None:
-            logger.error('Could not inflate %s, does not exist in storage', self)
-            return
-        self.cloudwanderer_metadata = updated_resource.cloudwanderer_metadata
-        self._set_attrs()
-
-    @property
-    def _clean_resource(self) -> dict:
-        return {
-            key: value
-            for key, value in self.metadata.items()
-            if not key.startswith('_')
-        }
-
-    def _set_attrs(self) -> None:
-        for key, value in self.cloudwanderer_metadata.resource_data.items():
-            setattr(
-                self,
-                xform_name(key),
-                value
-            )
-
-    def __repr__(self) -> str:
-        """Return a code representation of this resource."""
-        return str(
-            f"{self.__class__.__name__}("
-            f"urn={repr(self.urn)}, "
-            f"resource_data={self.cloudwanderer_metadata.resource_data}, "
-            f"resource_attributes={self.cloudwanderer_metadata.resource_attributes})"
-        )
-
-    def __str__(self) -> str:
-        """Return the string representation of this Resource."""
-        return repr(self)