+++ tests/integration/__init__.py
@@ -1,0 +1,3 @@
+import pytest
+pytest.register_assert_rewrite('tests.integration.storage_write_generic')
+pytest.register_assert_rewrite('tests.integration.storage_read_generic')