+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 0.9.0
+
+- Added `get_secondary_attribute` to `CloudWandererResource`
+- Added secondary attributes to `CloudWandererResource` if they don't clash.
+- Added `is_inflated` to `CloudWandererResource`
+
+- Storage Standardisation improvements
+    - Standardised storage connector write tests
+    - Deleted duplicate read tests
+    - Added `write_resource_attribute` to `BaseStorageConnector`
+    - Added `write_resource_attribute` to `MemoryStorageConnector`
+
+
 # 0.8.0
 
 - Added support for multiple storage connectors